[ML] Converting the anomaly details functions to typescript
